# GeneratedBySwagger/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Support your university with our API"
  version: "1.0.0"
  title: "PSIDI API"
host: "api.example.com"
basePath: "/v1"
schemes:
- "https"
produces:
- "application/json"
paths:
...
  /Utilizadores:
    get:
      tags:
      - "Utilizadores"
      summary: "Utilizadores registados no sistema."
      description: "Este Endpoint retorna todos os utilizadores incritos neste sistema,\
        \ podendo receber parametros que funcionaram como campos de pesquisa.\n"
      operationId: "utilizadoresGET"
      parameters:
      - name: "id_utilizador"
        in: "query"
        description: "Identificador do utilizador."
        required: false
        type: "string"
      - name: "nome"
        in: "query"
        description: "Nome do utilizador."
        required: false
        type: "string"
      responses:
        200:
          description: "Uma lista de utilizadores"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Utilizador"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Utilizadores"
    post:
      tags:
      - "Utilizadores"
      summary: "Inserção de um novo utilizador."
      description: "Este Endpoint regista novos utilizadores no sistema.\n"
      operationId: "utilizadoresPOST"
      parameters:
      - name: "id_utilizador"
        in: "query"
        description: "Identificador do novo utilizador."
        required: true
        type: "string"
      - name: "nome"
        in: "query"
        description: "Nome do novo utilizador."
        required: true
        type: "string"
      - name: "password"
        in: "query"
        description: "Password do novo utilizador."
        required: true
        type: "string"
      responses:
        200:
          description: "Id do utilizador registado."
          schema:
            type: "number"
        default:
          description: "Unexpected error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Utilizadores"
...
definitions:
...
  Utilizador:
    type: "object"
    properties:
      id_utilizador:
        type: "string"
        description: "Identificador do utilizador."
      nome:
        type: "string"
        description: "Nome do utilizador."
      password:
        type: "string"
        description: "Password do utilizador."
  Error:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      message:
        type: "string"
      fields:
        type: "string"